Fix signed integer overflow in Heretic FLOATRANDZ spawn height
P_SpawnMobj's FLOATRANDZ branch picks a random spawn height with
`space * P_Random()`, where space is a fixed_t. Any room taller than
about 436 units overflows int, which is undefined behaviour in C.

Vanilla Heretic wrapped, so the wrap is the behaviour demos depend on --
but nothing obliges the optimizer to preserve it. With LTO enabled,
clang acts on the assumption that it cannot happen and five Heretic
demos in the sync suite desync; building with -fwrapv makes them pass
again.

Do the multiply in unsigned, where wraparound is defined, and convert
back before the arithmetic shift so the result is bit-identical to what
the wrapping build produced.

Found with -fsanitize=signed-integer-overflow. Verified against the sync
suite (55 examples) and the heretic suite (1050 examples), with LTO on
and off.

+++ prboom2/src/p_mobj.c
@@ -1773,7 +1773,7 @@ mobj_t* P_SpawnMobj(fixed_t x,fixed_t y,fixed_t z,mobjtype_t type)
     {
       space -= 40 * FRACUNIT;
       mobj->z =
-        ((space * P_Random(pr_heretic)) >> 8) + mobj->floorz + 40 * FRACUNIT;
+        ((int)((unsigned) space * (unsigned) P_Random(pr_heretic)) >> 8) + mobj->floorz + 40 * FRACUNIT;
     }
     else
     {
